Updated preprocessing
I changed the code to analyze all the func files, not just the study files.
I attempted to switch the order of motion correction & slice timing correction, but I'm not sure how successful I was. The code runs, but the plots look identical, so I'm not sure. This is all I had time for today after our meeting.